+++ src/DataSources/Query/DateFilter.php
@@ -2,6 +2,8 @@
 
 namespace Notion\DataSources\Query;
 
+use DateTimeImmutable;
+use Notion\Common\Date;
 use stdClass;
 
 /** @psalm-immutable */
@@ -34,7 +36,7 @@ final readonly class DateFilter implements Filter, Condition
         private string $propertyType,
         private string $propertyName,
         private Operator $operator,
-        private string|bool|array|stdClass $value,
+        private DateTimeImmutable|bool|stdClass $value,
     ) {
         if (!in_array($operator, self::VALID_OPERATORS)) {
             throw new \Exception("Invalid operator");
@@ -86,7 +88,7 @@ final readonly class DateFilter implements Filter, Condition
         return $this->operator;
     }
 
-    public function value(): string|bool|array|stdClass
+    public function value(): DateTimeImmutable|bool|stdClass
     {
         return $this->value;
     }
@@ -94,31 +96,34 @@ final readonly class DateFilter implements Filter, Condition
     public function toArray(): array
     {
         $type = $this->propertyType === self::TYPE_PROPERTY ? "date" : $this->propertyName;
+        $value = $this->value instanceof DateTimeImmutable
+            ? $this->value->format(Date::FORMAT)
+            : $this->value;
 
         return [
             $this->propertyType() => $this->propertyName,
             $type => [
-                $this->operator->value => $this->value
+                $this->operator->value => $value,
             ],
         ];
     }
 
-    public function equals(string $value): self
+    public function equals(DateTimeImmutable $value): self
     {
         return new self($this->propertyType, $this->propertyName, Operator::Equals, $value);
     }
 
-    public function before(string $value): self
+    public function before(DateTimeImmutable $value): self
     {
         return new self($this->propertyType, $this->propertyName, Operator::Before, $value);
     }
 
-    public function after(string $value): self
+    public function after(DateTimeImmutable $value): self
     {
         return new self($this->propertyType, $this->propertyName, Operator::After, $value);
     }
 
-    public function onOrBefore(string $value): self
+    public function onOrBefore(DateTimeImmutable $value): self
     {
         return new self($this->propertyType, $this->propertyName, Operator::OnOrBefore, $value);
     }
@@ -133,7 +138,7 @@ final readonly class DateFilter implements Filter, Condition
         return new self($this->propertyType, $this->propertyName, Operator::IsNotEmpty, true);
     }
 
-    public function onOrAfter(string $value): self
+    public function onOrAfter(DateTimeImmutable $value): self
     {
         return new self($this->propertyType, $this->propertyName, Operator::OnOrAfter, $value);
     }